% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyCDP.R
\name{backtrace}
\alias{backtrace}
\title{backtrace: find the optimum path from the result of tinyCDP}
\usage{
backtrace(dp, pos)
}
\arguments{
\item{dp:}{result of DP maching calculated by tinyCDP()}

\item{pos:}{the detection position}
}
\value{
the optimum path:
}
\description{
backtrace: find the optimum path from the result of tinyCDP
}
