% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyCDP.R
\name{detect.peaks}
\alias{detect.peaks}
\title{detect.peaks: detect peaks from the vector}
\usage{
detect.peaks(values, thres, pwidth, type = "min")
}
\arguments{
\item{values}{a vector}

\item{thres}{the detection threshold}

\item{pwidth}{the inhibition width}

\item{type=c("min", "max")}{detection type}
}
\value{
vector of detection positions
}
\description{
detect.peaks: detect peaks from the vector
}
