% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyDP.R
\name{euclid2}
\alias{euclid2}
\title{euclid2: Squared Euclidean distance}
\usage{
euclid2(x, y)
}
\arguments{
\item{x, y}{A vector}
}
\value{
squared euclidean distance between x and y
}
\description{
euclid2: Squared Euclidean distance
}
