% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyDP.R
\name{eucrid2}
\alias{eucrid2}
\title{eucrid2: Squared Eucridean distance}
\usage{
eucrid2(x, y)
}
\arguments{
\item{x, y}{A vector}
}
\value{
squared euclidean distance between x and y
}
\description{
eucrid2: Squared Eucridean distance
}
