% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyDP.R
\name{manhattan}
\alias{manhattan}
\title{manhattan: Manhattan distance}
\usage{
manhattan(x, y)
}
\arguments{
\item{x, y}{A vector}
}
\value{
Manhattan (L1) distance between x and y
}
\description{
manhattan: Manhattan distance
}
