% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyDP.R
\name{negcos}
\alias{negcos}
\title{negcos: Negative cosine distance}
\usage{
negcos(x, y)
}
\arguments{
\item{x, y}{A vector}
}
\value{
Manhattan (L1) distance between x and y
}
\description{
negcos: Negative cosine distance
}
