% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyCDP.R
\name{tinyCDP}
\alias{tinyCDP}
\title{tinyCDP: the continuous DP matching}
\usage{
tinyCDP(d, extend_x = TRUE, extend_y = TRUE)
}
\arguments{
\item{d:}{the distance matrix (nrow(d) is the length of the database signal, ncol(d) is the length of the query signal)}
}
\value{
A list contains three elements:
 xsize: length (number of rows) of x
 ysize: length (number of rows) of y
 g: xsize*ysize matrix of the accumulated distance
 d: the distance matrix
 bp: the backpointer matrix
}
\description{
tinyCDP: the continuous DP matching
}
