% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyCDP.R
\name{tinyCDP.sqr}
\alias{tinyCDP.sqr}
\title{tinyCDP.sqr: the continuous DP matching using squared Euclidean distance}
\usage{
tinyCDP.sqr(x, y)
}
\arguments{
\item{x:}{a matrix of the input signal}

\item{y:}{a matrix of the signal to be detected}
}
\value{
A list contains three elements:
 xsize: length (number of rows) of x
 ysize: length (number of rows) of y
 g: xsize*ysize matrix of the accumulated distance
 d: the distance matrix
 bp: the backpointer matrix
}
\description{
tinyCDP.sqr: the continuous DP matching using squared Euclidean distance
}
