% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyDP.R
\name{tinyDP}
\alias{tinyDP}
\title{tinyDP: A fast dynamic time warping}
\usage{
tinyDP(x, y, window = 50, type = 1, distfunc = eucrid2)
}
\arguments{
\item{x}{A matrix}

\item{y}{A matrix}

\item{window}{The window width with which the DTW is calculated,}

\item{type}{If type==1, the DTW has not restriction on the length difference between x and y. If type==2, the length of the one matrix should be more than the half and less than the twice of the other matrix.}

\item{distfunc}{The distance function}
}
\value{
A list contains three elements:
 xsize: length (number of rows) of x
 ysize: length (number of rows) of y
 opt: n times 2 matrix, where a row contains the corresponding indices of x and y
 score: the matching distance
}
\description{
tinyDP: A fast dynamic time warping
}
